refactored quant modules
